This is a group collaboration for modelling the popular messaging app Telegram using a UML tool called USE
This is a group collaboration for modelling the popular messaging app Telegram using a UML tool called USE, in this folder you will find different diagrams related to the telegram system such as a class diagram, object diagram, sequence diagram and a state machine.  Provided also, is some documentation which provides an overall summary of the tasks done when creating this model of the application.